requirements.txt
# ── Core Framework ─────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9         # multipart/form-data upload support
aiofiles==24.1.0                # async file I/O for streaming upload

# ── Rate Limiting ───────────────────────────────────────────────────────────
slowapi==0.1.9

# ── PDF Processing ──────────────────────────────────────────────────────────
PyMuPDF==1.24.10                # fitz — digital text extraction
pdf2image==1.17.0               # PDF → PIL images (requires poppler system dep)
pdfplumber==0.11.4              # table extraction fallback

# ── OCR ─────────────────────────────────────────────────────────────────────
# PaddleOCR requires paddlepaddle installed separately:
#   CPU:  pip install paddlepaddle
#   GPU:  pip install paddlepaddle-gpu
paddleocr==2.8.1

# ── Image Processing ─────────────────────────────────────────────────────────
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

# ── Table Extraction ─────────────────────────────────────────────────────────
camelot-py[cv]==0.11.0          # requires ghostscript system dep
# tabula-py==2.9.3              # alt table extractor (Java required)

# ── Async Task Queue ─────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.8

# ── Logging ──────────────────────────────────────────────────────────────────
structlog==24.4.0

# ── Validation / NLP ─────────────────────────────────────────────────────────
langdetect==1.0.9               # language detection

# ── Optional: Layout Parser (ML-based layout detection) ─────────────────────
# layoutparser[layoutmodels,tesseract]==0.3.4
# detectron2 — install separately from: https://github.com/facebookresearch/detectron2

# ── Dev / Testing ────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.2                   # async test client for FastAPI
pytest-cov==5.0.0
